Renamed a package in the org.eclipse.smarthome.core.thing bundle. See openhab/static-code-analysis#100
