+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/internal/CommunicationManager.java
@@ -35,13 +35,13 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.events.ChannelTriggeredEvent;
 import org.eclipse.smarthome.core.thing.events.ThingEventFactory;
 import org.eclipse.smarthome.core.thing.internal.profiles.DefaultMasterProfile;
+import org.eclipse.smarthome.core.thing.internal.profiles.Profile;
+import org.eclipse.smarthome.core.thing.internal.profiles.ProfileFactory;
 import org.eclipse.smarthome.core.thing.internal.profiles.RawButtonToggleProfile;
+import org.eclipse.smarthome.core.thing.internal.profiles.StateProfile;
+import org.eclipse.smarthome.core.thing.internal.profiles.TriggerProfile;
 import org.eclipse.smarthome.core.thing.link.ItemChannelLink;
 import org.eclipse.smarthome.core.thing.link.ItemChannelLinkRegistry;
-import org.eclipse.smarthome.core.thing.profiles.Profile;
-import org.eclipse.smarthome.core.thing.profiles.ProfileFactory;
-import org.eclipse.smarthome.core.thing.profiles.StateProfile;
-import org.eclipse.smarthome.core.thing.profiles.TriggerProfile;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.State;
 import org.osgi.service.component.annotations.Component;
