+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/internal/profiles/TriggerProfile.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.core.thing.profiles;
+package org.eclipse.smarthome.core.thing.internal.profiles;
 
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.smarthome.core.events.EventPublisher;
